split resumableupload into separate files
