+++ frontend/javascripts/libs/resumable_upload/resumable_upload.ts
@@ -1,5 +1,9 @@
 // Inspired by Resumable.js (https://github.com/23/resumable.js)
 
+import type { ResumableChunk } from "./resumable_chunk";
+import { ResumableFile } from "./resumable_file";
+import { formatSize, generateUniqueIdentifier } from "./resumable_shared";
+
 /**
  * Configuration options for the Resumable instance.
  */
@@ -201,627 +205,6 @@ function isFileSystemFileEntry(
   return entry != null && entry.isFile;
 }
 
-// Helper functions
-const helpers = {
-  stopEvent(e: Event): void {
-    e.stopPropagation();
-    e.preventDefault();
-  },
-
-  generateUniqueIdentifier(file: File, _event?: Event): string {
-    const relativePath =
-      (file as any).webkitRelativePath || (file as any).relativePath || file.name;
-    const size = file.size;
-
-    return size + "-" + relativePath.replace(/[^0-9a-zA-Z_-]/gim, "");
-  },
-
-  formatSize(size: number | undefined): string {
-    if (size === undefined) {
-      return "n/a";
-    }
-    if (size < 1024) {
-      return size + " bytes";
-    } else if (size < 1024 * 1024) {
-      return (size / 1024.0).toFixed(0) + " KB";
-    } else if (size < 1024 * 1024 * 1024) {
-      return (size / 1024.0 / 1024.0).toFixed(1) + " MB";
-    } else {
-      return (size / 1024.0 / 1024.0 / 1024.0).toFixed(1) + " GB";
-    }
-  },
-
-  getTargetURI(resumable: ResumableUpload, request: string, params: Record<string, any>): string {
-    let target = resumable.getOpt("target");
-    const testTarget = resumable.getOpt("testTarget");
-
-    if (request === "test" && testTarget !== null) {
-      target = testTarget === "/" ? resumable.getOpt("target") : testTarget;
-    }
-
-    if (typeof target === "function") {
-      return target(params);
-    }
-
-    const url = new URL(target as string, window.location.origin);
-    Object.entries(params).forEach(([key, value]) => {
-      url.searchParams.append(key, value);
-    });
-
-    return url.toString();
-  },
-};
-
-/**
- * Represents a single chunk of a file to be uploaded.
- * @param preprocessState 0 = unprocessed, 1 = processing, 2 = finished
- */
-export class ResumableChunk {
-  opts: Partial<ConfigurationHash> = {};
-  resumableObj: ResumableUpload;
-  fileObj: ResumableFile;
-  fileObjSize: number;
-  fileObjType: string;
-  offset: number;
-  callback: (event: "progress" | "success" | "error" | "retry", message?: string) => void;
-  lastProgressCallback: Date;
-  tested = false;
-  retries = 0;
-  pendingRetry = false;
-  preprocessState: 0 | 1 | 2 = 0; // 0 = unprocessed, 1 = processing, 2 = finished
-  markComplete = false;
-  startByte: number;
-  endByte: number;
-
-  private abortController: AbortController | null = null;
-  private _status: "pending" | "uploading" | "success" | "error" = "pending";
-  private _message: string = "";
-
-  constructor(
-    resumableObj: ResumableUpload,
-    fileObj: ResumableFile,
-    offset: number,
-    callback: (event: "progress" | "success" | "error" | "retry", message?: string) => void,
-  ) {
-    this.resumableObj = resumableObj;
-    this.fileObj = fileObj;
-    this.fileObjSize = fileObj.size;
-    this.fileObjType = fileObj.file.type;
-    this.offset = offset;
-    this.callback = callback;
-    this.lastProgressCallback = new Date();
-
-    const chunkSize = this.getOpt("chunkSize") as number;
-    this.startByte = this.offset * chunkSize;
-    this.endByte = Math.min(this.fileObjSize, (this.offset + 1) * chunkSize);
-
-    if (this.fileObjSize - this.endByte < chunkSize && !this.getOpt("forceChunkSize")) {
-      this.endByte = this.fileObjSize;
-    }
-  }
-
-  getOpt<T extends keyof ConfigurationHash>(key: T): Required<ConfigurationHash>[T] {
-    if (this.opts[key] !== undefined) {
-      return this.opts[key] as Required<ConfigurationHash>[T];
-    }
-    return this.fileObj.getOpt(key);
-  }
-
-  async test(): Promise<void> {
-    this.abortController = new AbortController();
-
-    let customQueryParameters = this.getOpt("query");
-
-    if (typeof customQueryParameters === "function") {
-      customQueryParameters = customQueryParameters(this.fileObj, this);
-    }
-
-    const queryParams = Object.assign({}, customQueryParameters || {});
-
-    const extraParams: Partial<Record<string, any>> = {
-      [this.getOpt("chunkNumberParameterName")]: this.offset + 1,
-      [this.getOpt("chunkSizeParameterName")]: this.getOpt("chunkSize"),
-      [this.getOpt("currentChunkSizeParameterName")]: this.endByte - this.startByte,
-      [this.getOpt("totalSizeParameterName")]: this.fileObjSize,
-      [this.getOpt("typeParameterName")]: this.fileObjType,
-      [this.getOpt("identifierParameterName")]: this.fileObj.uniqueIdentifier,
-      [this.getOpt("fileNameParameterName")]: this.fileObj.fileName,
-      [this.getOpt("relativePathParameterName")]: this.fileObj.relativePath,
-      [this.getOpt("totalChunksParameterName")]: this.fileObj.chunks.length,
-    };
-
-    const targetUrl = helpers.getTargetURI(this.resumableObj, "test", {
-      ...queryParams,
-      ...extraParams,
-    });
-
-    let customHeaders = this.getOpt("headers");
-    if (typeof customHeaders === "function") {
-      customHeaders = customHeaders(this.fileObj, this);
-    }
-
-    const fetchTimeout = this.getOpt("fetchTimeout");
-    let hasTimedOut = false;
-    const timeoutId =
-      fetchTimeout > 0
-        ? setTimeout(() => {
-            hasTimedOut = true;
-            this.abortController?.abort();
-          }, fetchTimeout)
-        : null;
-
-    try {
-      const response = await fetch(targetUrl, {
-        method: this.getOpt("testMethod") as string,
-        headers: customHeaders,
-        signal: this.abortController.signal,
-        credentials: this.getOpt("withCredentials") ? "include" : "same-origin",
-      });
-      if (timeoutId != null) clearTimeout(timeoutId);
-      this.tested = true;
-
-      // Status 200: chunk already exists on server
-      // Status 204: chunk does not exist on server, please upload
-      if (response.ok && response.status !== 204) {
-        this._message = await response.text();
-        this.callback("success", this._message);
-        this.markComplete = true;
-        this.resumableObj.uploadNextChunk();
-      } else {
-        this.send();
-      }
-    } catch (error: any) {
-      if (timeoutId != null) clearTimeout(timeoutId);
-      if (error.name === "AbortError" && !hasTimedOut) return;
-
-      this.tested = true;
-      this.send();
-    }
-  }
-
-  preprocessFinished(): void {
-    this.preprocessState = 2;
-    this.send();
-  }
-
-  // send() uploads the actual data in a POST call
-  async send(): Promise<void> {
-    const preprocess = this.getOpt("preprocess");
-    if (typeof preprocess === "function") {
-      switch (this.preprocessState) {
-        case 0:
-          this.preprocessState = 1;
-          await Promise.resolve(preprocess(this));
-          return;
-        case 1:
-          return;
-        case 2:
-          break;
-      }
-    }
-
-    if (this.getOpt("testChunks") && !this.tested) {
-      this.test();
-      return;
-    }
-
-    this.abortController = new AbortController();
-    this._status = "uploading";
-    this.pendingRetry = false;
-    this.callback("progress");
-
-    const standardQueryParameters: Partial<Record<string, any>> = {
-      [this.getOpt("chunkNumberParameterName")]: this.offset + 1,
-      [this.getOpt("chunkSizeParameterName")]: this.getOpt("chunkSize"),
-      [this.getOpt("currentChunkSizeParameterName")]: this.endByte - this.startByte,
-      [this.getOpt("totalSizeParameterName")]: this.fileObjSize,
-      [this.getOpt("typeParameterName")]: this.fileObjType,
-      [this.getOpt("identifierParameterName")]: this.fileObj.uniqueIdentifier,
-      [this.getOpt("fileNameParameterName")]: this.fileObj.fileName,
-      [this.getOpt("relativePathParameterName")]: this.fileObj.relativePath,
-      [this.getOpt("totalChunksParameterName")]: this.fileObj.chunks.length,
-    };
-
-    let customQueryParameters = this.getOpt("query");
-    if (typeof customQueryParameters === "function") {
-      customQueryParameters = customQueryParameters(this.fileObj, this);
-    }
-    const queryParameters = Object.assign(standardQueryParameters, customQueryParameters || {});
-
-    const bytes = this.fileObj.file.slice(
-      this.startByte,
-      this.endByte,
-      this.getOpt("setChunkTypeFromFile") ? this.fileObj.file.type : "",
-    );
-
-    let data: FormData | Blob | string;
-
-    let customHeaders = this.getOpt("headers");
-    if (typeof customHeaders === "function") {
-      customHeaders = customHeaders(this.fileObj, this);
-    }
-    const headers: Record<string, string> = Object.assign({}, customHeaders);
-
-    if (this.getOpt("method") === "octet") {
-      data = bytes;
-      headers["Content-Type"] = "application/octet-stream";
-    } else {
-      const formData = new FormData();
-      // Add data from the query options
-      Object.entries(queryParameters).forEach(([k, v]) => {
-        formData.append(k, v);
-      });
-
-      if (this.getOpt("chunkFormat") === "blob") {
-        formData.append(this.getOpt("fileParameterName"), bytes, this.fileObj.fileName);
-        data = formData;
-      } else {
-        // chunkFormat == base64
-        const readPromise = new Promise<string>((resolve) => {
-          const fr = new FileReader();
-          fr.onload = () => resolve(fr.result as string);
-          fr.readAsDataURL(bytes);
-        });
-        const base64Data = await readPromise;
-        formData.append(this.getOpt("fileParameterName"), base64Data);
-        data = formData;
-      }
-    }
-
-    const targetUrl = helpers.getTargetURI(this.resumableObj, "upload", queryParameters);
-    const fetchTimeout = this.getOpt("fetchTimeout");
-    let hasTimedOut = false;
-    const timeoutId =
-      fetchTimeout > 0
-        ? setTimeout(() => {
-            hasTimedOut = true;
-            this.abortController?.abort();
-          }, fetchTimeout)
-        : null;
-
-    try {
-      const response = await fetch(targetUrl, {
-        method: this.getOpt("uploadMethod") as string,
-        headers,
-        body: data,
-        signal: this.abortController.signal,
-        credentials: this.getOpt("withCredentials") ? "include" : "same-origin",
-      });
-      if (timeoutId != null) clearTimeout(timeoutId);
-
-      if (response.ok) {
-        this._status = "success";
-        this._message = await response.text();
-        this.callback("success", this._message);
-        this.resumableObj.uploadNextChunk();
-      } else if (
-        this.getOpt("permanentErrors").includes(response.status) ||
-        this.retries >= this.getOpt("maxChunkRetries")
-      ) {
-        this._status = "error";
-        this._message = await response.text();
-        this.callback("error", this._message);
-        this.resumableObj.uploadNextChunk();
-      } else {
-        throw new Error(`Server responded with ${response.status}`);
-      }
-    } catch (error: any) {
-      if (timeoutId != null) clearTimeout(timeoutId);
-      if (error.name === "AbortError" && !hasTimedOut) {
-        return;
-      }
-
-      this.callback("retry", hasTimedOut ? "Timeout" : error.message);
-      this.abort();
-      this.retries++;
-
-      const retryInterval = this.getOpt("chunkRetryInterval");
-      if (retryInterval != null) {
-        this.pendingRetry = true;
-        setTimeout(() => this.send(), retryInterval);
-      } else {
-        this.send();
-      }
-    }
-  }
-
-  abort(): void {
-    if (this.abortController) {
-      this.abortController.abort();
-      this.abortController = null;
-    }
-    this._status = "pending";
-  }
-
-  status(): "pending" | "uploading" | "success" | "error" {
-    // if pending retry then that's effectively the same as actively uploading,
-    // there might just be a slight delay before the retry starts
-    if (this.pendingRetry) return "uploading";
-    if (this.markComplete) return "success";
-    return this._status;
-  }
-
-  message(): string {
-    return this._message;
-  }
-
-  progress(relative = false): number {
-    // NOTE: the fetch API doesn't provide a way to get the progress of the upload so we need to track the uploaded bytes ourselves
-    let factor = relative ? (this.endByte - this.startByte) / this.fileObjSize : 1;
-    if (this.pendingRetry) return 0;
-    if (this.status() === "uploading" && !this.markComplete) factor *= 0.95;
-
-    const s = this.status();
-    switch (s) {
-      case "success":
-      case "error":
-        return 1 * factor;
-      case "pending":
-      default:
-        return 0;
-    }
-  }
-}
-
-/**
- * Represents a file to be uploaded.
- */
-export class ResumableFile {
-  opts: Partial<ConfigurationHash> = {};
-  /**
-   * A back-reference to the parent `Resumable` object.
-   */
-  resumableObj: ResumableUpload;
-  /**
-   * The correlating HTML5 `File` object.
-   */
-  file: File;
-  /**
-   * The name of the file.
-   */
-  fileName: string;
-  /**
-   * Size in bytes of the file.
-   */
-  size: number;
-  /**
-   * The relative path to the file (defaults to file name if relative path doesn't exist)
-   */
-  relativePath: string;
-  /**
-   * A unique identifier assigned to this file object. This value is included in uploads to the server for reference, but can also be used in CSS classes etc when building your upload UI.
-   */
-  uniqueIdentifier: string;
-  /**
-   * An array of `ResumableChunk` items.
-   */
-  chunks: ResumableChunk[] = [];
-  container: EventTarget | null = null;
-  preprocessState: 0 | 1 | 2 = 0; // 0 = unprocessed, 1 = processing, 2 = finished
-  private _prevProgress = 0;
-  private _pause = false;
-  private _error: boolean;
-
-  constructor(resumableObj: ResumableUpload, file: File, uniqueIdentifier: string) {
-    this.resumableObj = resumableObj;
-    this.file = file;
-    this.fileName = file.name;
-    this.size = file.size;
-    this.relativePath =
-      (file as any).relativePath || (file as any).webkitRelativePath || this.fileName;
-    this.uniqueIdentifier = uniqueIdentifier;
-    this._error = uniqueIdentifier === undefined;
-
-    this.resumableObj.dispatchFromDetail({ type: "chunkingStart", file: this });
-    this.bootstrap();
-  }
-
-  hasError() {
-    return this._error;
-  }
-
-  getOpt<T extends keyof ConfigurationHash>(key: T): Required<ConfigurationHash>[T] {
-    if (this.opts[key] !== undefined) {
-      return this.opts[key] as Required<ConfigurationHash>[T];
-    }
-    return this.resumableObj.getOpt(key);
-  }
-
-  // Callback when something happens within the chunk
-  private chunkEvent(event: "progress" | "success" | "error" | "retry", message?: string): void {
-    switch (event) {
-      case "progress":
-        this.resumableObj.dispatchFromDetail({ type: "fileProgress", file: this, message });
-        break;
-      case "error":
-        this.abort();
-        this._error = true;
-        this.chunks = [];
-        this.resumableObj.dispatchFromDetail({ type: "fileError", file: this, message });
-        break;
-      case "success":
-        if (this._error) return;
-        this.resumableObj.dispatchFromDetail({ type: "fileProgress", file: this, message });
-        if (this.isComplete()) {
-          this.resumableObj.dispatchFromDetail({ type: "fileSuccess", file: this, message });
-        }
-        break;
-      case "retry":
-        this.resumableObj.dispatchFromDetail({ type: "fileRetry", file: this });
-        break;
-    }
-  }
-
-  /**
-   * Abort uploading the file.
-   */
-  abort(): void {
-    let abortCount = 0;
-    for (const chunk of this.chunks) {
-      if (chunk.status() === "uploading") {
-        chunk.abort();
-        abortCount++;
-      }
-    }
-    if (abortCount > 0) {
-      this.resumableObj.dispatchFromDetail({ type: "fileProgress", file: this });
-    }
-  }
-
-  /**
-   * Abort uploading the file and delete it from the list of files to upload (reset).
-   */
-  cancel(): void {
-    const chunks = this.chunks;
-    this.chunks = [];
-    for (const chunk of chunks) {
-      if (chunk.status() === "uploading") {
-        chunk.abort();
-        this.resumableObj.uploadNextChunk();
-      }
-    }
-    this.resumableObj.removeFile(this);
-    this.resumableObj.dispatchFromDetail({ type: "fileProgress", file: this });
-  }
-
-  /**
-   * Retry uploading the file.
-   */
-  retry(): void {
-    this.bootstrap();
-    const handler = (e: Event) => {
-      const detail = (e as CustomEvent).detail;
-      if (detail.file === this) {
-        this.resumableObj.upload();
-        this.resumableObj.removeEventListener("chunkingComplete", handler);
-      }
-    };
-    this.resumableObj.addEventListener("chunkingComplete", handler);
-  }
-
-  /**
-   * Rebuild the state of a `ResumableFile` object, including reassigning chunks and XMLHttpRequest instances.
-   */
-  bootstrap(): void {
-    this.abort();
-    this._error = false;
-    this.chunks = [];
-    this._prevProgress = 0;
-
-    const round = this.getOpt("forceChunkSize") ? Math.ceil : Math.floor;
-    const maxOffset = Math.max(round(this.file.size / this.getOpt("chunkSize")), 1);
-
-    // Rebuild stack of chunks from file
-    for (let offset = 0; offset < maxOffset; offset++) {
-      this.chunks.push(
-        new ResumableChunk(this.resumableObj, this, offset, (event, message) =>
-          this.chunkEvent(event, message),
-        ),
-      );
-      this.resumableObj.dispatchFromDetail({
-        type: "chunkingProgress",
-        file: this,
-        message: (offset / maxOffset).toString(),
-      });
-    }
-
-    setTimeout(() => {
-      this.resumableObj.dispatchFromDetail({ type: "chunkingComplete", file: this });
-    }, 0);
-  }
-
-  /**
-   * Returns a float between 0 and 1 indicating the current upload progress of the file. If `relative` is `true`, the value is returned relative to all files in the Resumable.js instance.
-   */
-  progress(): number {
-    if (this._error) return 1;
-
-    let returnValue = 0;
-    let error = false;
-    for (const chunk of this.chunks) {
-      if (chunk.status() === "error") error = true;
-      returnValue += chunk.progress(true); // get chunk progress relative to entire file
-    }
-    returnValue = error ? 1 : returnValue > 0.99999 ? 1 : returnValue;
-    returnValue = Math.max(this._prevProgress, returnValue); // We don't want to lose percentages when an upload is paused
-    this._prevProgress = returnValue;
-    return returnValue;
-  }
-
-  /**
-   * Returns a boolean indicating whether file chunks is uploading.
-   */
-  isUploading(): boolean {
-    return this.chunks.some((chunk) => chunk.status() === "uploading");
-  }
-
-  /**
-   * Returns a boolean indicating whether the file has completed uploading and received a server response.
-   */
-  isComplete(): boolean {
-    if (this.preprocessState === 1) return false;
-
-    return !this.chunks.some(
-      (chunk) =>
-        chunk.status() === "uploading" ||
-        chunk.status() === "pending" ||
-        chunk.preprocessState === 1,
-    );
-  }
-
-  pause(pause?: boolean): void {
-    this._pause = pause === undefined ? !this._pause : pause;
-  }
-
-  isPaused(): boolean {
-    return this._pause;
-  }
-
-  preprocessFinished(): void {
-    this.preprocessState = 2;
-    this.upload();
-  }
-
-  upload(): boolean {
-    if (this.isPaused() === false) {
-      const preprocess = this.getOpt("preprocessFile");
-      if (typeof preprocess === "function") {
-        switch (this.preprocessState) {
-          case 0: {
-            this.preprocessState = 1;
-            const result = preprocess(this);
-            if (result instanceof Promise) result.catch(() => {});
-            return true;
-          }
-          case 1:
-            return true;
-          case 2:
-            break;
-        }
-      }
-
-      for (const chunk of this.chunks) {
-        if (chunk.status() === "pending" && chunk.preprocessState !== 1) {
-          chunk.send();
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  /**
-   * starts upload from the next chunk number while marking all previous chunks complete. Must be called before upload() method
-   */
-  markChunksCompleted(chunkNumber: number): void {
-    if (!this.chunks || this.chunks.length <= chunkNumber) return;
-
-    for (let num = 0; num < chunkNumber; num++) {
-      if (this.chunks[num]) this.chunks[num].markComplete = true;
-    }
-  }
-}
-
 /**
  * Main class for managing resumable uploads.
  * Implements EventTarget via composition to ensure compatibility.
@@ -891,13 +274,13 @@ export class ResumableUpload implements EventTarget {
       minFileSize: 1,
       minFileSizeErrorCallback: (file: File) => {
         console.error(
-          `${file.name} is too small, please upload files larger than ${helpers.formatSize(this.getOpt("minFileSize"))}.`,
+          `${file.name} is too small, please upload files larger than ${formatSize(this.getOpt("minFileSize"))}.`,
         );
       },
       maxFileSize: Number.POSITIVE_INFINITY,
       maxFileSizeErrorCallback: (file: File) => {
         console.error(
-          `${file.name} is too large, please upload files less than ${helpers.formatSize(this.getOpt("maxFileSize"))}.`,
+          `${file.name} is too large, please upload files less than ${formatSize(this.getOpt("maxFileSize"))}.`,
         );
       },
       fileType: [],
@@ -1199,7 +582,7 @@ export class ResumableUpload implements EventTarget {
       };
 
       // directories have size == 0
-      const uniqueIdentifier = helpers.generateUniqueIdentifier(file, event);
+      const uniqueIdentifier = generateUniqueIdentifier(file, event);
       const customGenerator = this.getOpt("generateUniqueIdentifier");
 
       if (typeof customGenerator === "function") {
