+++ frontend/javascripts/test/libs/resumable-upload-use-cases.spec.ts
@@ -1,8 +1,8 @@
+import { ResumableUpload, type ResumableUploadEvent } from "libs/resumable_upload/resumable_upload";
 import { sleep } from "libs/utils";
 import { HttpResponse, http } from "msw";
 import { setupServer } from "msw/node";
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
-import { ResumableUpload, type ResumableUploadEvent } from "../../libs/resumable-upload";
 import { ResumableBackendMock } from "../helpers/resumable_backend_mock";
 
 describe("Resumable Use Cases (WebKnossos Patterns)", () => {
